Initial commit of unumpy
An initial commit of unumpy. I use the numpy.vectorize function to
vectorize each of the umath functions so that they may be applied to
numpy arrays.

Note that, at the moment, the abs function is not included, as this
conflicts with a python keyword. I’m sure there is a better way to
handle this?